+++ app/src/main/java/com/gl/habitalarm/ui/createoredit/HabitCreateOrEditActivity.kt
@@ -1,5 +1,7 @@
-package com.gl.habitalarm.ui.create
+package com.gl.habitalarm.ui.createoredit
 
+import android.content.Context
+import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.MenuItem
@@ -7,14 +9,14 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import com.gl.habitalarm.R
-import com.gl.habitalarm.databinding.ActivityCreateBinding
+import com.gl.habitalarm.databinding.ActivityHabitCreateOrEditBinding
 import dagger.hilt.android.AndroidEntryPoint
 
-private const val TAG = "CreateActivity"
+private const val TAG = "HabitCreateOrEditActivity"
 
 @AndroidEntryPoint
-class CreateActivity : AppCompatActivity() {
-    private val mViewModel: CreateViewModel by viewModels()
+class HabitCreateOrEditActivity : AppCompatActivity() {
+    private val mViewModel: HabitCreateOrEditViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         Log.d(TAG, "onCreate(): called")
@@ -31,8 +33,8 @@ class CreateActivity : AppCompatActivity() {
                 }
         )
 
-        val binding: ActivityCreateBinding =
-            DataBindingUtil.setContentView(this, R.layout.activity_create)
+        val binding: ActivityHabitCreateOrEditBinding =
+            DataBindingUtil.setContentView(this, R.layout.activity_habit_create_or_edit)
         binding.viewModel = mViewModel
         binding.lifecycleOwner = this
     }
@@ -40,5 +42,14 @@ class CreateActivity : AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         finishAfterTransition()
         return true
+    }
+
+    companion object {
+        fun createIntent(packageContext: Context?, habitId: Long): Intent? {
+            Log.d(TAG, "createIntent(): called with habitId $habitId")
+
+            return Intent(packageContext, HabitCreateOrEditActivity::class.java)
+                    .putExtra(HabitCreateOrEditViewModel.INTENT_HABIT_ID, habitId)
+        }
     }
 }
