+++ app/src/main/res/layout/activity_habit_create_or_edit.xml
@@ -3,7 +3,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.create.CreateActivity">
+    tools:context=".ui.createoredit.HabitCreateOrEditActivity">
 
     <data>
         <import type="androidx.databinding.adapters.TimePickerBindingAdapter" />
@@ -12,7 +12,7 @@
 
         <variable
             name="viewModel"
-            type="com.gl.habitalarm.ui.create.CreateViewModel" />
+            type="com.gl.habitalarm.ui.createoredit.HabitCreateOrEditViewModel" />
     </data>
 
     <ScrollView
